Tried to add a code of writing data into tree.json - for the first look it works fine
